Dialogue code reorganization
