+++ Content/UI/Dialogue/GuideCamp.xml
@@ -33,7 +33,7 @@
 			<Text time="60">Maybe look for a [caravan merchant] or someone who can confirm the rumors.</Text>
 		</Dialogue>
 		<Options>
-			<Option link="exit" icon="chest" rewardIndex="0" action="quest_complete">Turn In</Option>
+			<Option link="exit" icon="chest" rewardIndex="monster_reward" action="quest_complete">Turn In</Option>
 		</Options>
 	</DialogueNode>
 
@@ -44,16 +44,16 @@
 			<Text time="240">Regardless, you'll have to get some better gear before you try anything like that. I don't get out much these days so why don't you borrow these for now. The rest you'll have to find on your own.</Text>
 		</Dialogue>
 		<Options>
-			<Option link="exit" icon="chest" rewardIndex="1" action="quest_complete">Turn In</Option>
+			<Option link="exit" icon="chest" rewardIndex="biome_reward" action="quest_complete">Turn In</Option>
 		</Options>
 	</DialogueNode>
 
 	<DialogueNode id="quest_complete_build">
 		<Dialogue>
 			<Text time="180">Ah, settling down are we? You think you could find a place for these in that fancy new home of ours? Things start to get heavy after so many years.</Text>
-			<Text time="60">Don't get excited, it won't be anything too fancy</Text>
 		</Dialogue>
 		<Options>
+			<Option link="quest_complete_build_end">Don't get excited, it won't be anything too fancy</Option>
 			<Option link="quest_complete_build_end">I'll make them the best storage space you've ever seen</Option>
 		</Options>
 	</DialogueNode>
@@ -63,7 +63,7 @@
 			<Text time="60">I'm sure whatever you build it will be great. Stay safe out there!</Text>
 		</Dialogue>
 		<Options>
-			<Option link="exit" icon="chest" rewardIndex="2" action="quest_complete">Turn In</Option>
+			<Option link="exit" icon="chest" rewardIndex="build_reward" action="quest_complete">Turn In</Option>
 		</Options>
 	</DialogueNode>
 </Dialogue>
